implementando features de mostrar o valor na tela home e validacoes

+++ app/src/main/java/br/com/dionataferraz/vendas/data/local/AccountEntity.kt
@@ -8,7 +8,7 @@ import java.util.*
 data class AccountEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
-    val accountBalance: Int,
+    val accountBalance: Double,
     val type: Type,
     val date: String
 )
